tests : refactor vocab tests
ggml-ci
